Setup the infrastructure necessary to run tests
I have borrowed heavily from the MIM Makefile to put together
everything necessary to run the local tests and the ejabberd
integration tests. Not everything works quite right yet: some of the
integration tests fail and I haven't tested the dialyzer targets.

Also, I tweaked the compile and deps tasks to try and speed up
compile, which is going to be run frequently but was quite slow.
Since the deps are unlikely to change, we compile them once when
they are fetched and do not try to recompile them every time compile
is run. This brings the compile task down to about 8s from 30s when
all of the deps are downloaded and compiled and there is nothing
else to do. Still too long, but much better than before.