Recreate the reltool release using relx
This is intended to recreate the reltool release as closely as possible.
The relx config is simpler, but the main reason for doing it is that it
will make things better in the long term. The current release is very
cluttered, but with some TLC it will be svelte and streamlined and relx
will help with that.

NOTE: The release generation currently fails with the alarms dependency.
alarms has configuration elements in the .app file which are not correct
according to the docs, but which some versions of reltool will let past.
relx is apparently more strict.