Greatly reduce the number of vars in vars.config
The large number of variables in vars.config made the deployment overly
complicated. Reduce the number of interpolated variables to the bare
minimum necessary to build the primary release package and test nodes.
Largely this is done by hard coding information that really doesn't need
to change.

Also, run ejabberd.cfg through Emacs' indenter to clean it up and sort
the *vars.config files by use. All of the vars files should have the
same set of variables now, reducing the likelihood of a bad interpolation.