Store the SCRAM password in a single database field
We can serialize the SCRAM record into a string and store it in the
database as such. This is how the Riak and ODBC auth providers do it.
This is nice because we can avoid leaking ejabberd SCRAM details into the
wocky_user model and keep our concerns nice and separated.