Get eunit, CT, dialyzer and xref working through rebar
The Makefile is pretty bare at this point. Also, create CT test stubs for
the wocky code.

Note that the test/ejabberd directory was causing problems with rebar so
I removed it. The integration test automation still works, it is run from
the MongooseIM directory rather than being symlinked into the top level
of the project.