Refactor the Cassandra startup and configuration
The primary purpose of this change is to allow
wocky to run in a unit/functional test suite
without starting the full ejabberd application. To
accomplish this, there are three significant
changes:

1. Application startup has been changed such that
ejabberd isn’t always automatically started. You
can specify whether you want ejabberd to start
by setting a flag in the wocky environment.

2. You can configure wocky as before by putting
settings in ejabberd.cfg (with one minor change),
but you can also put the configuration settings in
the wocky.app file, in sys.config or in a file
passed in a -config option on the command line.

3. Worker pool management has been delegated to
the Pooler library. This simplifies the management
code and allows us to dynamically scale up workers
with load (to a specified maximum).