Make sure the C compiler can find openssl
Recent versions of OS X do not come with openssl
installed. Apple is using their own SSL library
and we have to install openssl via Homebrew. The
drawback is that Homebrew does not put openssl
into a common directory to avoid clashes.

I had been setting this in my shell manually and
finally got tired of it.