Add basic Vagrantfile
With this in place, you can run `vagrant up` to spin up a new virtual machine with Erlang installed.

It is not quite ready for prime time, though. It doesn't have Cassandra installed and we probably want to have a clean copy of the code instead of sharing the host's copy.

When this is done, we should be able to start a local VM or an EC2 instance with the same base OS and provisioning.