Add wocky_app:start/1 which takes config
rebar3 changes the way that CT tests are run with the result that we can't pass an application config file to the Erlang VM like we could with earlier versions. We can work around this by passing the config we want to use to start/1 and injecting it into the application environment. It is kinda kludgey, but it works.

The wocky_user_SUITE tests now all run successfully. The only catch is that you have to load the test schema (in priv/test-schema.cql) manually once before running the tests or wocky will fail to start. At some point I will automate that.