Create a kludgey script to wrap rebar
Right now it just sets the environment so that the test config is loaded at the appropriate time. I am sure it will be needed for other ugly hacks down the road.

This is so fracking ugly. It boggles the mind that the rebar authors never considered that one might want to load app configuration when running tests.