Create a functional test for ejabberd_auth_wocky
This is based on the test suite for ejabberd_auth_http in the MIM sources. We cover all the cases they do, with the exception of the passoword management functions when SCRAM is disabled. I am not sure there is great value in testing that case.

The coverage for ejabberd_auth_wocky isn't very high, but there is a ton of boilerplate code in that module that would be very low-value to test.