Change course on the roster implementation
Implementing a backend for EJD's mod_roster is going to be insufficient for our long-term needs. Instead, we want to consider implmenting our own roster module that takes advantage of the C* architecture and allows us to add custom fields to the roster.

This is a first step in this direction. The mod_wocky_roster module is a refactored and reorganized version of the EJD roster module. It still works the same way, with the exception that code related to disabling roster versioning has been removed.