Fix the semantics for `wocky_db:batch_query`
The `batch_query` function was originally based off of an example in the cqerl docs and it ran a single query with multiple data sets. This turns out to be exactly wrong. In C*, batches are meant to be used to keep multiple different tables in sync and not to process a large chunk of data efficiently.

Change the semantics so that `batch_query` takes a list of query, values pairs and does the right thing. Since the use case of "run the same query with different data sets" is a legitimate one, also add a new function `multi_query` to do that.