Drop null values from result set rows
Rows in the result set are just proplists. When a column doesn't have a value in the db, it comes back as `{column_name, null}`. This causes some problems when we try to use the normal Erlang property list functions. For example, we would normally expect to get `undefined` from `proplists:get_value/2` when a value doesn't exist. The old code returned `null`. The old code also fails to work as expected with `proplists:get_value/3`.