Rationalize user-related modules
This change is based on a team discussion about what is in the local part of the JID. Make sure that we are treating the local part of the JID as the timeuuid associated with a user. The username is separate and generally is not used at this layer of the code.

Also, make user parameter names and ordering are consistent across modules.