First cut at a high level C* API
The idea here is to create a high level API for C* that will make existing db code simpler and allow us to do more schema manipulation in code. The latter should make both manual and automated testing easier.

The API as it is now represents about 70% of what we will ultimately need. I am shooting for the simplest thing that will meet the goals and not a full-featured API that covers all of what you can do through CQL. As a result, I am focusing only on generating CQL statements that we use and not covering features that we don't need now.