Define a baseline ejabberd.cfg
To get to the baseline config, I removed all of the settings for features that we aren't currently using. I also removed all of the comments for settings we aren't using and the explanatory comments for the settings that we are. All of our Wocky modules are enabled and configured.

The result is the minimum viable configuration, or pretty close to it. Our integration tests all pass using the config and the core ejabberd functionality seems to be intact. This won't be the final config we use, but it should provide a solid foundation for moving forward.

I have added a copy of the original file as `etc/ejabberd.default.cfg` to use as a reference.