Ensure that CT suites use the canonical config
I extracted the code necessary to boot EJD in the tests into a new `test_helper` module. Instead of each suite having its own copy of the config, they now refer to the one canonical version in etc.

The test configs were diverging from the production config, so we should try to use the production config for testing whenever possible.