Enable specifying the C* node in the app env
The cassandra_nodes setting in the cqerl environment seeds the connection pool, but that is it. You still need all of the connection settings when you get a new client. Which makes the cassandra_nodes setting...useless. The whole process of getting a new client and how the connection informtion is passed in cqerl needs to be rethought.