Use a cleaner/shorter version number
We are iterating so fast that the normal X.Y.Z version numbers do not seem useful. Instead use a version number of the form YY.MM.DD+N where N is the number of git commits on HEAD at the time of the build. This version number encodes useful information (the date of the build) and uses the number of commits to disambiguate multiple builds on the same day.