Temporary fix for CT privacy suite
This fixes two problems:

First, it adds the privacy module to test.cfg until we can reenable it in _global.cfg.

Second, privacy_SUITE was leaving privacy settings behind that were blocking the presence notifications and causing the cases in activity_SUITE to fail.

Neither of these solutions is ideal. This is a short-term fix to get the CT suite working again. Ideally, each test suite should clean up after itself OR each suite needs to clear the entire database at the beginning of the run.