Manage log levels better
ejabberd_app forces the loglevel for ALL Lager backends to info on startup. This is hardcoded and cannot be changed. In addition, if you set the 'loglevel' parameter in ejabberd.cfg, then all backends will be set to that level.

To address this, we don't set 'loglevel' in any of the config files. Second, after ejabberd starts, we reset the levels for any backends that need to be at anything other than 'info'.

This change also tweaks logging a bit. There are now 3 main log files:

* erlang.log: contains console output, set to info
* wocky.log: set to warning, rotated
* debug.log: set to debug, rotated