Add DB storage and retrieval of lookup limits
A counter is stored in the db. It is retrieved at the beginning of a query and stored at the end. The counter starts at 5000 and 1 is subtracted for each query performed. When the counter reaches 0 no more queries are performed. Since the counter counts down, it is referred to as "reductions" in the code, similar to the Erlang scheduler.