Return 404 earlier to avoid a crash in cqerl
When the path was wrong, user_file/1 was returning two empty strings for the user and file ID. These were getting eventually passed into cqerl as the value of timeuuid fields and causing cqerl to crash generating the query. Catch this condition early and return a 404 to avoid all of that.