Handle checking whether a blank username exists
mod_mam:filter_packet/1 looks up the 'To' user for every local packet that comes in. Many of these packets (such as disco and ping packets) are to the server and don't have a localpart of the JID. Cassandra returns an error if the string is empty, so we create a special case and avoid the trip to the database.