Redo the api for pulling all of the user data
Renamed `get_user_data` to `find_user` and the `get_user_by_foo` functions are combined into `find_user_by`.  Also, the `find_user_by` function now returns the full user data rather than just the user id and server.

The convention established here is that the functions that retrieve a single value for an existing user are prefixed with `get_`, but the functions that return all of the data for a user are prefixed with `find_`.