Avoid saving multiple entries for the same device id
We were seeing multiple push notifications being delivered to a user for each event. Part of the problem seems to be that the client is registering the same device id multiple times for different resource names. Change the tables so that the device id is in the key and resource name is not.