Stop pushing notifications after one succeeds
The client appears to be registering multiple device ids for a single physical device. Stop sending push notifications after one succeeds so that users won't get multiple notifications.

This isn't what we want long-term and only works now because we expect users to have a single device. This will need to be revisited before a wider production release.