Fix malformed xmlel records in integration tests
The new version of `exml` is less forgiving than the older versions. In particular, the `children` member of the `xmlel` record must be formatted properly: a shallow list of `xmlel` or `xmlcdata` records.

There were two common problems with the way `xmlel` records were being constructed in the test suite; either the `children` member was a single `xmlel` record instead of a list or the `children` member contained nested lists.