Experimentally raise the max_fsm_queue limit
We are seeing crashes as this limit is hit when sending offline messages or homestream notifications. It doesn't happen often, so I tend to think that a small increase might be sufficient to prevent it.

If the crash continues to happen with the higher limit, then we should set it back to the old value and investigate throttling message traffic when a client resumes a session.