Fix a badarith error during bot search
The lat and lon are being passed into the bot search as strings, so we need to convert them to floats.

The fix I provided is a bit Q+D. If the string representation is actually an integer, it will throw an error. We shouldn't be getting whole degrees, so it shouldn't be an issue; but it is something to be aware of.