Use Elixir to generate version numbers
Elixir 1.4.4 includes a more stringent SemVer module that will fail if any segment of a version number includes a leading "0". Unfortunately, the shell version of the version script did just this and caused the build to fail under 1.4.4. Using Elxir for this gives more control over the output and we can generate a number that SemVer will be happy with.