Always index the whole object
Previously, we would try to optimize the communication with Algolia and only send an update if one of the indexed fields had changed. Unfortunately, since both users and bots are partially created before being fully fleshed out, this could lead to some objects missing important fields in the Algolia index.

For now, we are not going to try to optimize that communication and just send the whole object to the indexer every time it changes. This will lead to unnecessary chattiness, but it will mean that we should not miss any important fields.