Fix a client authentication issue
The client was including text in the "domain" part of the SASL PLAIN authentication string. This was causing us to fail to authenticate. I have fixed it so that the authentication succeeds regardless of whether there is text in the "domain" bit and I have added some logging to help with debugging future issues.