Catch exceptions in the db event poller
There was an issue in prod where the default GenServer timeout was exceeded by `send_to.send()` causing the poller to crash. Theoretically, `source.recv()` and `send_to.send()` could raise any exception or crash. So, we wrap them in a try block and use `after` to guarantee the poll loop continues.