Fix dependency hell
Somewhere along the way our dependencies got a little out of whack. As long as you worked from the top of the umbrella, everything was fine. But if you tried to run mix in any of the app dirs, you would get dependency errors. Specifying some overrides fixes this annoyance.