Process batch location uploads in the ReST API
If a batch of locations are uploaded at once, they are all sent through geofence processing, but only the last one will be sent to live location sharing.

Note that the data points must be processed in order, so there will be some lag before the live shared location data is sent while the earlier data points are processed.