Get the revision to deploy from the repo
Instead of taking the revision to deploy on the command line, the deploy script now gets the revision from the repo. This means that you have to check out the version that you want to deploy. This enforces the rule that the deployment script should match the code being deployed.