Use User.tid() more often
A few rules for this change:

Replace User.id() with User.tid() in almost all cases since there is no real downside.

Replace User.t() with User.tid() in the following cases:

1. When the function only uses the ID
2. When the function only passes the value to another function that uses User.tid()

Functions that currently take a User.t() and use the fully hydrated object are generally left alone since it is usually better to hydrate the object higher in the call chain. I also changed a few functions that took only an ID but used a hydrated object to just take the hydrated object.