Correct a race condition in the client tests
The client tests are getting a GQL error from certain notifications. It appears that the error is due to a race condition where the "other user" in a notification is deleted in between the db callback firing and the notification being sent.

We resolve the issue here by loading the other user from the database earlier, and not sending the notification if the other user doesn't exist.